MainWindow UI optimizations
Got Style-happy since a lot of what was in the XAML was repeated over and over (especially with the bottom buttons).

Also made it to where the buttons are in a StackPanel, which makes them evenly spaced out; and I removed a lot of unnecessary resources that were included when I copied the templates for the items I was editing.